backend/requirements.txt
# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database - PostgreSQL
sqlalchemy>=2.0.25
psycopg[binary]>=3.1.18

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1  # Fixed version for passlib compatibility

# Data Validation
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0

# OCR & Document Processing
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.2.0
PyPDF2>=3.0.1

# NLP & NER
spacy>=3.7.2

# Machine Learning
scikit-learn>=1.4.0
pandas>=2.1.4
numpy>=1.26.3
joblib>=1.3.2

# Environment
python-dotenv>=1.0.0

# CORS
starlette>=0.35.1

